# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PartSignal API
  version: 0.1.1
  description: PartSignal MVP 契约。除登录外，写请求必须携带 X-CSRF-Token。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/geo-observations:
    get:
      tags: [observation]
      operationId: listGeoObservations
      responses:
        '200':
          description: 观测列表
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GeoObservationList'}
    post:
      tags: [observation]
      operationId: createGeoObservation
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GeoObservationCreate'}
      responses:
        '201':
          description: 已追加观测
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GeoObservation'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: partsignal_session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 32}
...
  schemas:
...
    RecommendationStatus: {type: string, enum: [NONE, CANDIDATE, RECOMMENDED]}
    AccuracyStatus: {type: string, enum: [ACCURATE, PARTIAL, INCORRECT, UNJUDGEABLE]}
    CitationSourceType: {type: string, enum: [OFFICIAL, EXTERNAL_COMPANY, OTHER]}
    GeoCitation:
      type: object
      additionalProperties: false
      required: [url, source_type]
      properties:
        url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        source_type: {$ref: '#/components/schemas/CitationSourceType'}
        publication_record_id: {type: [string, 'null'], format: uuid}
    GeoObservationCreate:
      type: object
      additionalProperties: false
      required: [query_topic_id, product_id, actual_prompt, model_name, tested_at, web_search_enabled, answer_summary, mentioned, recommendation, accuracy, citations, publication_record_ids, notes]
      properties:
        query_topic_id: {type: string, format: uuid}
        product_id: {type: string, format: uuid}
        actual_prompt: {type: string}
        model_name: {type: string}
        model_version: {type: [string, 'null']}
        tested_at: {type: string, format: date-time}
        web_search_enabled: {type: boolean}
        answer_summary: {type: string}
        mentioned: {type: boolean}
        recommendation: {$ref: '#/components/schemas/RecommendationStatus'}
        accuracy: {$ref: '#/components/schemas/AccuracyStatus'}
        citations: {type: array, items: {$ref: '#/components/schemas/GeoCitation'}}
        publication_record_ids: {type: array, uniqueItems: true, items: {type: string, format: uuid}}
        attachment_file_ids: {type: array, uniqueItems: true, items: {type: string, format: uuid}}
        notes: {type: string}
        supersedes_id: {type: [string, 'null'], format: uuid}
    GeoObservation:
      allOf:
        - {$ref: '#/components/schemas/GeoObservationCreate'}
        - type: object
          required: [id, tested_by, created_at]
          properties:
            id: {type: string, format: uuid}
            tested_by: {type: string, format: uuid}
            created_at: {type: string, format: date-time}
    GeoObservationList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/GeoObservation'}}
...